requirements.txt
django==4.2.3
djangorestframework==3.14.0
pytest-django==4.5.2
drf-yasg==1.21.6
djangorestframework-simplejwt==5.2.2
django-rest-passwordreset==1.3.0
coreapi==2.3.3
Pillow==10.0.0